% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean.model}
\alias{clean.model}
\title{Remove all files and folders created during build of pdmphmc model}
\usage{
clean.model(build.object, remove.all = FALSE, dry.run = FALSE)
}
\arguments{
\item{build.object}{model made using the build()-function.}
}
\description{
Remove all files and folders created made using the build()-function. Note: model will not work after this call
}
