% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{get_CPU_time}
\alias{get_CPU_time}
\title{Get CPU times from pdphmc output}
\usage{
get_CPU_time(object)
}
\arguments{
\item{object}{S4 class of type `run-output` typically generated by function `run`}
}
\description{
CPU times from pdphmc output
}
