% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{histogram.plot}
\alias{histogram.plot}
\title{Histogram plot from pdphmc output}
\usage{
histogram.plot(object, which.par = NULL)
}
\arguments{
\item{object}{S4 class of type `run-output` typically generated by function `run`}

\item{which.par}{character vector of parameter/generated names}
}
\description{
Histogram plot from pdphmc output
}
