% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{pair.plot}
\alias{pair.plot}
\title{pair plot from pdphmc output}
\usage{
pair.plot(object, which.par.x = NULL, which.par.y = NULL)
}
\arguments{
\item{object}{S4 class of type `run-output` typically generated by function `run`}

\item{which.par.x}{character vector of parameter/generated names on x-axis}

\item{which.par.y}{character vector of parameter/generated names on y-axis}
}
\description{
pair plot of pdmphmc output
}
