% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run pdmphmc model}
\usage{
run(
  build.object,
  samples = 2000L,
  warmupFrac = 0.5,
  Tmax = 10000,
  data = list(),
  control = pdmphmc_control(),
  chains = 4L,
  cores = 1L,
  seed = 1L,
  fixedMiDiag = list(),
  store.pars = NULL,
  clean = TRUE
)
}
\arguments{
\item{build.object}{model made using the build()-function.}

\item{samples}{How many equidistant samples to record.}

\item{warmupFrac}{Which proportion of trajectory is to be used for warmup (tuning and reaching stationary phase).}

\item{Tmax}{(time-)length of trajectories}

\item{data}{named list of data, with names and formats corresponding to the DATA_* statements in the model struct}

\item{control}{named list for setting tuning parameters (integrator tolerances etc)}

\item{chains}{(integer>0) How many trajectories should be run?}

\item{cores}{(0<integer<=chains) How many cores should be used? Based on foreach-package.}

\item{seed}{(integer) random number seed.}

\item{fixedMiDiag}{(named list) set the scaling/mass matrix diagonal to a fixed quantity (turn of adaption), e.g. if the model contains a parameter par, then list(par=2.0) fixes the scaling so that the posterior of par is assumed to have variance around 2.0}

\item{store.pars}{(character vector) which parameters are to be stored. If empty, then all parameters are stored. Generated quantites are always stored, and should not be passed using this argument}

\item{clean}{(logical) remove output files after they have been read into R? Mainly for debugging purposes}
}
\description{
Run a model made using the build()-function
}
