% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsityTools.R
\name{sparsity.tools}
\alias{sparsity.tools}
\title{optimizing the sparse cholesky factorization}
\usage{
sparsity.tools(object, spy = TRUE)
}
\arguments{
\item{object}{S4 class of type `run-output` typically generated by function `run`}
}
\description{
Methods for optimizing the sparse cholesky factorization
}
