% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDmatrix_utils.R
\name{SPDmatrix.asmatrix}
\alias{SPDmatrix.asmatrix}
\title{Transform a SPDmatrix representation vector to a SPD matrix}
\usage{
SPDmatrix.asmatrix(vec)
}
\arguments{
\item{vec}{a SPDmatrix vector representation to be transformed to a SPD matrix.}
}
\description{
Transform a SPDmatrix representation vector to a SPD matrix
}
