% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDmatrix_utils.R
\name{SPDmatrix.asvector}
\alias{SPDmatrix.asvector}
\title{Transform a SPD matrix to the vector representation SPDmatrix}
\usage{
SPDmatrix.asvector(mat)
}
\arguments{
\item{mat}{a SPD matrix to be transformed}
}
\description{
Transform a SPD matrix to the vector representation SPDmatrix
}
