% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{trace.plot}
\alias{trace.plot}
\title{trace plot from pdphmc output}
\usage{
trace.plot(object, which.par = NULL)
}
\arguments{
\item{object}{S4 class of type `run-output` typically generated by function `run`}

\item{which.par}{character vector of parameter/generated names}
}
\description{
trace plot from pdphmc output
}
